requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
argument-clinic==0.12
    # via -r requirements.in
click==8.1.3
    # via pip-tools
docstring-parser==0.14.1
    # via argument-clinic
imageio==2.19.3
    # via -r requirements.in
imageio-ffmpeg==0.4.7
    # via -r requirements.in
numpy==1.22.4
    # via imageio
pep517==0.12.0
    # via pip-tools
pillow==9.1.1
    # via
    #   -r requirements.in
    #   imageio
pip-tools==6.6.2
    # via -r requirements.in
tomli==2.0.1
    # via pep517
tqdm==4.64.0
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
